+++ examples/basic.js
@@ -1,6 +1,6 @@
 var api = require('../lib/ApiQuick').init(8080);
-api.addPackage('date', 
-	function(method, args, params) {
-		return {date: new Date().toUTCString()};
-	}
-);
+var endpoints = {};
+endpoints.date = function(method, args, params) {
+	return {date: new Date().toUTCString()};
+};
+api.addEndpoints(endpoints);