+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 + Tests for the full program
 + Disable auth by setting ```auth: false```
 + Fix dependency versions
++ Mark addPackage as deprecated, due to be removed v1.0.0
++ Implement addEndpoints which will replace addPackage
 
 ## 0.1.2 BETA (2016-05-17)